Struct wasmtime_runtime::component::ComponentInstance

#[repr(C)]
pub struct ComponentInstance { /* private fields */ }

Runtime representation of a component instance and all state necessary for the instance itself.

This type never exists by-value, but rather it’s always behind a pointer. The size of the allocation for ComponentInstance includes the trailing VMComponentContext which is variably sized based on the offsets contained within.

Implementations

impl ComponentInstance

pub unsafe fn from_vmctx<R>( vmctx: *mut VMComponentContext, f: impl FnOnce(&mut ComponentInstance) -> R ) -> R

Converts the vmctx provided into a ComponentInstance and runs the provided closure with that instance.

Unsafety

This is unsafe because vmctx cannot be guaranteed to be a valid pointer and it cannot be proven statically that it’s safe to get a mutable reference at this time to the instance from vmctx.

pub fn instance_flags( &self, instance: RuntimeComponentInstanceIndex ) -> InstanceFlags

Returns a pointer to the “may leave” flag for this instance specified for canonical lowering and lifting operations.

pub fn store(&self) -> *mut dyn Store

Returns the store that this component was created with.

pub fn runtime_memory(&self, idx: RuntimeMemoryIndex) -> *mut VMMemoryDefinition

Returns the runtime memory definition corresponding to the index of the memory provided.

This can only be called after idx has been initialized at runtime during the instantiation process of a component.

pub fn runtime_realloc(&self, idx: RuntimeReallocIndex) -> NonNull<VMFuncRef>

Returns the realloc pointer corresponding to the index provided.

This can only be called after idx has been initialized at runtime during the instantiation process of a component.

pub fn runtime_post_return( &self, idx: RuntimePostReturnIndex ) -> NonNull<VMFuncRef>

Returns the post-return pointer corresponding to the index provided.

This can only be called after idx has been initialized at runtime during the instantiation process of a component.

pub fn lowering(&self, idx: LoweredIndex) -> VMLowering

Returns the host information for the lowered function at the index specified.

This can only be called after idx has been initialized at runtime during the instantiation process of a component.

pub fn trampoline_func_ref(&self, idx: TrampolineIndex) -> NonNull<VMFuncRef>

Returns the core wasm funcref corresponding to the trampoline specified.

The returned function is suitable to pass directly to a wasm module instantiation and the function contains cranelift-compiled trampolines.

This can only be called after idx has been initialized at runtime during the instantiation process of a component.

pub fn set_runtime_memory( &mut self, idx: RuntimeMemoryIndex, ptr: *mut VMMemoryDefinition )

Stores the runtime memory pointer at the index specified.

This is intended to be called during the instantiation process of a component once a memory is available, which may not be until part-way through component instantiation.

Note that it should be a property of the component model that the ptr here is never needed prior to it being configured here in the instance.

pub fn set_runtime_realloc( &mut self, idx: RuntimeReallocIndex, ptr: NonNull<VMFuncRef> )

Same as set_runtime_memory but for realloc function pointers.

pub fn set_runtime_post_return( &mut self, idx: RuntimePostReturnIndex, ptr: NonNull<VMFuncRef> )

Same as set_runtime_memory but for post-return function pointers.

pub fn set_lowering(&mut self, idx: LoweredIndex, lowering: VMLowering)

Configures host runtime lowering information associated with imported f functions for the idx specified.

pub fn set_trampoline( &mut self, idx: TrampolineIndex, wasm_call: NonNull<VMWasmCallFunction>, native_call: NonNull<VMNativeCallFunction>, array_call: VMArrayCallFunction, type_index: VMSharedTypeIndex )

Same as set_lowering but for the resource.drop functions.

pub fn set_resource_destructor( &mut self, idx: ResourceIndex, dtor: Option<NonNull<VMFuncRef>> )

Configures the destructor for a resource at the idx specified.

This is required to be called for each resource as it’s defined within a component during the instantiation process.

pub fn resource_destructor( &self, idx: ResourceIndex ) -> Option<NonNull<VMFuncRef>>

Returns the destructor, if any, for idx.

This is only valid to call after set_resource_destructor, or typically after instantiation.

pub fn component(&self) -> &Component

Returns a reference to the component type information for this instance.

pub fn component_types(&self) -> &Arc<ComponentTypes>

Returns the type information that this instance is instantiated with.

pub fn realloc_func_ty(&self) -> &Arc<dyn Any + Send + Sync>

Get the canonical ABI’s realloc function’s runtime type.

pub fn resource_types(&self) -> &Arc<dyn Any + Send + Sync>

Returns a reference to the resource type information as a dyn Any.

Wasmtime is the one which then downcasts this to the appropriate type.

pub fn resource_owned_by_own_instance(&self, ty: TypeResourceTableIndex) -> bool

Returns whether the resource that ty points to is owned by the instance that ty correspond to.

This is used when lowering borrows to skip table management and instead thread through the underlying representation directly.

pub fn resource_new32( &mut self, resource: TypeResourceTableIndex, rep: u32 ) -> Result<u32>

Implementation of the resource.new intrinsic for i32 representations.

pub fn resource_rep32( &mut self, resource: TypeResourceTableIndex, idx: u32 ) -> Result<u32>

Implementation of the resource.rep intrinsic for i32 representations.

pub fn resource_drop( &mut self, resource: TypeResourceTableIndex, idx: u32 ) -> Result<Option<u32>>

Implementation of the resource.drop intrinsic.

pub fn component_resource_tables( &mut self ) -> &mut PrimaryMap<TypeResourceTableIndex, ResourceTable>

Returns the runtime state of resources associated with this component.

pub fn dtor_and_flags( &self, ty: TypeResourceTableIndex ) -> (Option<NonNull<VMFuncRef>>, Option<InstanceFlags>)

Returns the destructor and instance flags for the specified resource table type.

This will lookup the origin definition of the ty table and return the destructor/flags for that.