Struct wasmtime_runtime::component::ResourceTables

pub struct ResourceTables<'a> {
    pub tables: Option<&'a mut PrimaryMap<TypeResourceTableIndex, ResourceTable>>,
    pub host_table: Option<&'a mut ResourceTable>,
    pub calls: &'a mut CallContexts,
}

Contextual state necessary to perform resource-related operations.

This state a bit odd since it has a few optional bits, but the idea is that whenever this is constructed the bits required to perform operations are always Some. For example:

During lifting and lowering both tables and host_table are Some.

During wasm’s own intrinsics only tables is Some.

During embedder-invoked resource destruction calls only host_table is Some.

This is all packaged up into one state though to make it easier to operate on and to centralize handling of the state related to resources due to how critical it is for correctness.

Fields
tables: Option<&'a mut PrimaryMap<TypeResourceTableIndex, ResourceTable>>

Runtime state for all resources defined in a component.

This is required whenever a TypeResourceTableIndex is provided as it’s the lookup where that happens. Not present during embedder-originating operations though such as ResourceAny::resource_drop which won’t consult this table as it’s only operating over the host table.

host_table: Option<&'a mut ResourceTable>

Runtime state for resources currently owned by the host.

This is the single table used by the host stored within Store<T>. Host resources will point into this and effectively have the same semantics as-if they’re in-component resources. The major distinction though is that this is a heterogenous table instead of only containing a single type.

calls: &'a mut CallContexts

Scope information about calls actively in use to track information such as borrow counts.

Implementations

impl ResourceTables<'_>

pub fn resource_new( &mut self, ty: Option<TypeResourceTableIndex>, rep: u32 ) -> Result<u32>

Implementation of the resource.new canonical intrinsic.

Note that this is the same as resource_lower_own.

pub fn resource_rep( &mut self, ty: Option<TypeResourceTableIndex>, idx: u32 ) -> Result<u32>

Implementation of the resource.rep canonical intrinsic.

This one’s one of the simpler ones: “just get the rep please”

pub fn resource_drop( &mut self, ty: Option<TypeResourceTableIndex>, idx: u32 ) -> Result<Option<u32>>

Implementation of the resource.drop canonical intrinsic minus the actual invocation of the destructor.

This will drop the handle at the idx specified, removing it from the specified table. This operation can fail if:

The index is invalid.

The index points to an own resource which has active borrows.

Otherwise this will return Some(rep) if the destructor for rep needs to run. If None is returned then that means a borrow handle was removed and no destructor is necessary.

pub fn resource_lower_own( &mut self, ty: Option<TypeResourceTableIndex>, rep: u32 ) -> Result<u32>

Inserts a new “own” handle into the specified table.

This will insert the specified representation into the specified type table.

Note that this operation is infallible, and additionally that this is the same as resource_new implementation-wise.

This is an implementation of the canonical ABI lower_own function.

pub fn resource_lift_own( &mut self, ty: Option<TypeResourceTableIndex>, idx: u32 ) -> Result<u32>

Attempts to remove an “own” handle from the specified table and its index.

This operation will fail if idx is invalid, if it’s a borrow handle, or if the own handle has currently been “lent” as a borrow.

This is an implementation of the canonical ABI lift_own function.

pub fn resource_lift_borrow( &mut self, ty: Option<TypeResourceTableIndex>, idx: u32 ) -> Result<u32>

Extracts the underlying resource representation by lifting a “borrow” from the tables.

This primarily employs dynamic tracking when a borrow is created from an “own” handle to ensure that the “own” handle isn’t dropped while the borrow is active and additionally that when the current call scope returns the lend operation is undone.

This is an implementation of the canonical ABI lift_borrow function.

pub fn resource_lower_borrow( &mut self, ty: Option<TypeResourceTableIndex>, rep: u32 ) -> Result<u32>

Records a new borrow resource with the given representation within the current call scope.

This requires that a call scope is active. Additionally the number of active borrows in the latest scope will be increased and must be decreased through a future use of resource_drop before the current call scope exits.

This some of the implementation of the canonical ABI lower_borrow function. The other half of this implementation is located on VMComponentContext which handles the special case of avoiding borrow tracking entirely.

pub fn enter_call(&mut self)

Enters a new calling context, starting a fresh count of borrows and such.

pub fn exit_call(&mut self) -> Result<()>

Exits the previously pushed calling context.

This requires all information to be available within this ResourceTables and is only called during lowering/lifting operations at this time.