Module wasmtime_runtime::libcalls::raw

Raw functions which are actually called from compiled code.

Invocation of a builtin currently looks like:

A wasm function calls a cranelift-compiled trampoline that’s generated once-per-builtin.

The cranelift-compiled trampoline performs any necessary actions to exit wasm, such as dealing with fp/pc/etc.

The cranelift-compiled trampoline loads a function pointer from an array stored in VMContext That function pointer is defined in this module.

This module runs, handling things like catch_unwind and Result and such.

This module delegates to the outer module (this file) which has the actual implementation.

Functions

activations_table_insert_with_gc⚠

check_free⚠

check_load⚠

check_malloc⚠

check_store⚠

data_drop⚠

drop_externref⚠

elem_drop⚠

externref_global_get⚠

externref_global_set⚠

free_start⚠

malloc_start⚠

memory32_grow⚠

memory_atomic_notify⚠

memory_atomic_wait32⚠

memory_atomic_wait64⚠

memory_copy⚠

memory_fill⚠

memory_init⚠

new_epoch⚠

out_of_gas⚠

ref_func⚠

table_copy⚠

table_fill_externref⚠

table_fill_func_ref⚠

table_get_lazy_init_func_ref⚠

table_grow_externref⚠

table_grow_func_ref⚠

table_init⚠

update_mem_size⚠

update_stack_pointer⚠