wasmtime_runtime::libcalls
Module relocs

This module contains functions which are used for resolving relocations at runtime if necessary.

These functions are not used by default and currently the only platform they’re used for is on x86_64 when SIMD is disabled and then SSE features are further disabled. In these configurations Cranelift isn’t allowed to use native CPU instructions so it falls back to libcalls and we rely on the Rust standard library generally for implementing these.

Functions

ceilf32

ceilf64

floorf32

floorf64

fmaf32

fmaf64

nearestf32

nearestf64

truncf32

truncf64