Struct wasmtime_runtime::VMArrayCallHostFuncContext

#[repr(C)]
pub struct VMArrayCallHostFuncContext { /* private fields */ }

The VM*Context for array-call host functions.

Its magic field must always be wasmtime_environ::VM_ARRAY_CALL_HOST_FUNC_MAGIC, and this is how you can determine whether a VM*Context is a VMArrayCallHostFuncContext versus a different kind of context.

Implementations

impl VMArrayCallHostFuncContext

pub unsafe fn new( func_ref: VMFuncRef, host_state: Box<dyn Any + Send + Sync> ) -> StoreBox<VMArrayCallHostFuncContext>

Create the context for the given host function.

Safety

The host_func must be a pointer to a host (not Wasm) function and it must be Send and Sync.

pub fn host_state(&self) -> &(dyn Any + Send + Sync)

Get the host state for this host function context.

pub fn func_ref(&self) -> &VMFuncRef

Get this context’s VMFuncRef.

pub unsafe fn from_opaque( opaque: *mut VMOpaqueContext ) -> *mut VMArrayCallHostFuncContext

Helper function to cast between context types using a debug assertion to protect against some mistakes.