Module wasmtime_runtime::component

Runtime support for the component model in Wasmtime

Currently this runtime support includes a VMComponentContext which is similar in purpose to VMContext. The context is read from cranelift-generated trampolines when entering the host from a wasm module. Eventually it’s intended that module-to-module calls, which would be cranelift-compiled adapters, will use this VMComponentContext as well.

Structs

ComponentInstance

Runtime representation of a component instance and all state necessary for the instance itself.

InstanceFlags

OwnedComponentInstance

An owned version of ComponentInstance which is akin to Box<ComponentInstance>.

VMComponentContext

This is a marker type to represent the underlying allocation of a VMComponentContext.

VMLowering

Structure describing a lowered host function stored within a VMComponentContext per-lowering.

Type Definitions

VMLoweringCallee

Type signature for host-defined trampolines that are called from WebAssembly.