Struct wasmtime_runtime::component::ComponentInstance

#[repr(C)]
pub struct ComponentInstance { /* private fields */ }

Runtime representation of a component instance and all state necessary for the instance itself.

This type never exists by-value, but rather it’s always behind a pointer. The size of the allocation for ComponentInstance includes the trailing VMComponentContext which is variably sized based on the offsets contained within.

Implementations

impl ComponentInstance

pub fn instance_flags(
&self,
instance: RuntimeComponentInstanceIndex
) -> InstanceFlags

Returns a pointer to the “may leave” flag for this instance specified for canonical lowering and lifting operations.

pub fn store(&self) -> *mut dyn Store

Returns the store that this component was created with.

pub fn runtime_memory(&self, idx: RuntimeMemoryIndex) -> *mut VMMemoryDefinition

Returns the runtime memory definition corresponding to the index of the memory provided.

This can only be called after idx has been initialized at runtime during the instantiation process of a component.

pub fn runtime_realloc(
&self,
idx: RuntimeReallocIndex
) -> NonNull<VMCallerCheckedAnyfunc>

Returns the realloc pointer corresponding to the index provided.

This can only be called after idx has been initialized at runtime during the instantiation process of a component.

pub fn runtime_post_return(
&self,
idx: RuntimePostReturnIndex
) -> NonNull<VMCallerCheckedAnyfunc>

Returns the post-return pointer corresponding to the index provided.

This can only be called after idx has been initialized at runtime during the instantiation process of a component.

pub fn lowering(&self, idx: LoweredIndex) -> VMLowering

Returns the host information for the lowered function at the index specified.

This can only be called after idx has been initialized at runtime during the instantiation process of a component.

pub fn lowering_anyfunc(
&self,
idx: LoweredIndex
) -> NonNull<VMCallerCheckedAnyfunc>

Returns the core wasm function pointer corresponding to the lowering index specified.

The returned function is suitable to pass directly to a wasm module instantiation and the function is a cranelift-compiled trampoline.

This can only be called after idx has been initialized at runtime during the instantiation process of a component.

pub fn always_trap_anyfunc(
&self,
idx: RuntimeAlwaysTrapIndex
) -> NonNull<VMCallerCheckedAnyfunc>

Same as lowering_anyfunc except for the functions that always trap.

pub fn transcoder_anyfunc(
&self,
idx: RuntimeTranscoderIndex
) -> NonNull<VMCallerCheckedAnyfunc>

Same as lowering_anyfunc except for the transcoding functions.

pub fn set_runtime_memory(
&mut self,
idx: RuntimeMemoryIndex,
ptr: *mut VMMemoryDefinition
)

Stores the runtime memory pointer at the index specified.

This is intended to be called during the instantiation process of a component once a memory is available, which may not be until part-way through component instantiation.

Note that it should be a property of the component model that the ptr here is never needed prior to it being configured here in the instance.

pub fn set_runtime_realloc(
&mut self,
idx: RuntimeReallocIndex,
ptr: NonNull<VMCallerCheckedAnyfunc>
)

Same as set_runtime_memory but for realloc function pointers.

pub fn set_runtime_post_return(
&mut self,
idx: RuntimePostReturnIndex,
ptr: NonNull<VMCallerCheckedAnyfunc>
)

Same as set_runtime_memory but for post-return function pointers.

pub fn set_lowering(
&mut self,
idx: LoweredIndex,
lowering: VMLowering,
anyfunc_func_ptr: NonNull<VMFunctionBody>,
anyfunc_type_index: VMSharedSignatureIndex
)

Configures a lowered host function with all the pieces necessary.

idx - the index that’s being configured

lowering - the host-related closure information to get invoked when the lowering is called.

anyfunc_func_ptr - the cranelift-compiled trampoline which will read the VMComponentContext and invoke lowering provided. This function pointer will be passed to wasm if wasm needs to instantiate something.

anyfunc_type_index - the signature index for the core wasm type registered within the engine already.

pub fn set_always_trap(
&mut self,
idx: RuntimeAlwaysTrapIndex,
func_ptr: NonNull<VMFunctionBody>,
type_index: VMSharedSignatureIndex
)

Same as set_lowering but for the “always trap” functions.

pub fn set_transcoder(
&mut self,
idx: RuntimeTranscoderIndex,
func_ptr: NonNull<VMFunctionBody>,
type_index: VMSharedSignatureIndex
)

Same as set_lowering but for the transcoder functions.