Type Alias wasmtime_runtime::component::VMLoweringCallee

pub type VMLoweringCallee = extern "C" fn(vmctx: *mut VMOpaqueContext, data: *mut u8, ty: TypeFuncIndex, flags: InstanceFlags, opt_memory: *mut VMMemoryDefinition, opt_realloc: *mut VMFuncRef, string_encoding: StringEncoding, args_and_results: *mut MaybeUninit<ValRaw>, nargs_and_results: usize);

Type signature for host-defined trampolines that are called from WebAssembly.

This function signature is invoked from a cranelift-compiled trampoline that adapts from the core wasm System-V ABI into the ABI provided here:

vmctx - this is the first argument to the wasm import, and should always end up being a VMComponentContext.

data - this is the data pointer associated with the VMLowering for which this function pointer was registered.

ty - the type index, relative to the tables in vmctx, that is the type of the function being called.

flags - the component flags for may_enter/leave corresponding to the component instance that the lowering happened within.

opt_memory - this nullable pointer represents the memory configuration option for the canonical ABI options.

opt_realloc - this nullable pointer represents the realloc configuration option for the canonical ABI options.

string_encoding - this is the configured string encoding for the canonical ABI this lowering corresponds to.

args_and_results - pointer to stack-allocated space in the caller where all the arguments are stored as well as where the results will be written to. The size and initialized bytes of this depends on the core wasm type signature that this callee corresponds to.

nargs_and_results - the size, in units of ValRaw, of args_and_results.