Enum wasmtime_runtime::Trap

pub enum Trap {
    User {
        error: Error,
        backtrace: Option<Backtrace>,
    },
    Jit {
        pc: usize,
        backtrace: Option<Backtrace>,
    },
    Wasm {
        trap_code: TrapCode,
        backtrace: Option<Backtrace>,
    },
    OOM {
        backtrace: Option<Backtrace>,
    },
}

Stores trace message with backtrace.

Variants

User

Fields

error: Error

The user-provided error

backtrace: Option<Backtrace>

Native stack backtrace at the time the trap occurred

A user-raised trap through raise_user_trap.

Jit

Fields

pc: usize

The program counter in JIT code where this trap happened.

backtrace: Option<Backtrace>

Native stack backtrace at the time the trap occurred

A trap raised from jit code

Wasm

Fields

trap_code: TrapCode

Code of the trap.

backtrace: Option<Backtrace>

Native stack backtrace at the time the trap occurred

A trap raised from a wasm libcall

OOM

Fields

backtrace: Option<Backtrace>

Native stack backtrace at the time the OOM occurred

A trap indicating that the runtime was unable to allocate sufficient memory.

Implementations

impl Trap

pub fn wasm(trap_code: TrapCode) -> Self

Construct a new Wasm trap with the given trap code.

Internally saves a backtrace when passed across a setjmp boundary, if the engine is configured to save backtraces.

pub fn user(error: Error) -> Self

Construct a new Wasm trap from a user Error.

Internally saves a backtrace when passed across a setjmp boundary, if the engine is configured to save backtraces.

pub fn oom() -> Self

Construct a new OOM trap.

Internally saves a backtrace when passed across a setjmp boundary, if the engine is configured to save backtraces.