Enum wasmtime_runtime::Trap

pub enum Trap {
    User(Box<dyn Error + Send + Sync>),
    Jit {
        pc: usize,
        backtrace: Backtrace,
        maybe_interrupted: bool,
    },
    Wasm {
        trap_code: TrapCode,
        backtrace: Backtrace,
    },
    OOM {
        backtrace: Backtrace,
    },
}

Stores trace message with backtrace.

Variants

User(Box<dyn Error + Send + Sync>)

A user-raised trap through raise_user_trap.

Jit

A trap raised from jit code

Fields of Jit

pc: usize

The program counter in JIT code where this trap happened.

backtrace: Backtrace

Native stack backtrace at the time the trap occurred

maybe_interrupted: bool

An indicator for whether this may have been a trap generated from an interrupt, used for switching what would otherwise be a stack overflow trap to be an interrupt trap.

Wasm

A trap raised from a wasm libcall

Fields of Wasm

trap_code: TrapCode

Code of the trap.

backtrace: Backtrace

Native stack backtrace at the time the trap occurred

OOM

A trap indicating that the runtime was unable to allocate sufficient memory.

Fields of OOM

backtrace: Backtrace

Native stack backtrace at the time the OOM occurred

Implementations

impl Trap

pub fn wasm(trap_code: TrapCode) -> Self

Construct a new Wasm trap with the given source location and trap code.

Internally saves a backtrace when constructed.

pub fn oom() -> Self

Construct a new OOM trap with the given source location and trap code.

Internally saves a backtrace when constructed.