Enum wasmtime_runtime::Trap

pub enum Trap {
    User(Error),
    Jit {
        pc: usize,
        backtrace: Backtrace,
        maybe_interrupted: bool,
    },
    Wasm {
        trap_code: TrapCode,
        backtrace: Backtrace,
    },
    OOM {
        backtrace: Backtrace,
    },
}

Stores trace message with backtrace.

Variants

User(Error)

A user-raised trap through raise_user_trap.

Jit

Fields

pc: usize

The program counter in JIT code where this trap happened.

backtrace: Backtrace

Native stack backtrace at the time the trap occurred

maybe_interrupted: bool

An indicator for whether this may have been a trap generated from an interrupt, used for switching what would otherwise be a stack overflow trap to be an interrupt trap.

A trap raised from jit code

Wasm

Fields

trap_code: TrapCode

Code of the trap.

backtrace: Backtrace

Native stack backtrace at the time the trap occurred

A trap raised from a wasm libcall

OOM

Fields

backtrace: Backtrace

Native stack backtrace at the time the OOM occurred

A trap indicating that the runtime was unable to allocate sufficient memory.

Implementations

impl Trap

pub fn wasm(trap_code: TrapCode) -> Self

Construct a new Wasm trap with the given source location and trap code.

Internally saves a backtrace when constructed.

pub fn oom() -> Self

Construct a new OOM trap with the given source location and trap code.

Internally saves a backtrace when constructed.