wasmtime_runtime

Function gc

pub unsafe fn gc(
    module_info_lookup: &dyn ModuleInfoLookup,
    externref_activations_table: &mut VMExternRefActivationsTable,
)

Perform garbage collection of VMExternRefs.

Unsafety

You must have called VMExternRefActivationsTable::set_stack_canary for at least the oldest host–>Wasm stack frame transition on this thread’s stack (it is idempotent to call it more than once) and keep its return value alive across the duration of that host–>Wasm call.

Additionally, you must have registered the stack maps for every Wasm module that has frames on the stack with the given stack_maps_registry.