wasmtime_runtime::libcalls

Module trampolines

Actually public trampolines which are used by the runtime as the entrypoint for libcalls.

Note that the trampolines here are actually defined in inline assembly right now to ensure that the fp/sp on exit are recorded for backtraces to work properly.

Functions

activations_table_insert_with_gc⚠

data_drop⚠

drop_externref⚠

elem_drop⚠

externref_global_get⚠

externref_global_set⚠

memory32_grow⚠

memory_atomic_notify⚠

memory_atomic_wait32⚠

memory_atomic_wait64⚠

memory_copy⚠

memory_fill⚠

memory_init⚠

new_epoch⚠

out_of_gas⚠

ref_func⚠

table_copy⚠

table_fill_externref⚠

table_fill_funcref⚠

table_get_lazy_init_funcref⚠

table_grow_externref⚠

table_grow_funcref⚠

table_init⚠