Struct wasmtime_runtime::InstanceHandle

pub struct InstanceHandle { /* private fields */ }

A handle holding an Instance of a WebAssembly module.

Implementations

impl InstanceHandle

pub unsafe fn from_vmctx(vmctx: *mut VMContext) -> Self

Create a new InstanceHandle pointing at the instance pointed to by the given VMContext pointer.

Safety

This is unsafe because it doesn’t work on just any VMContext, it must be a VMContext allocated as part of an Instance.

pub fn vmctx(&self) -> &VMContext

Return a reference to the vmctx used by compiled wasm code.

pub fn vmctx_ptr(&self) -> *mut VMContext

Return a raw pointer to the vmctx used by compiled wasm code.

pub fn module(&self) -> &Module

Return a reference to a module.

pub fn lookup_by_declaration(&self, export: &EntityIndex) -> Export

Lookup an export with the given export declaration.

pub fn exports(&self) -> Iter<'_, String, EntityIndex>

Return an iterator over the exports of this instance.

Specifically, it provides access to the key-value pairs, where the keys are export names, and the values are export declarations which can be resolved lookup_by_declaration.

pub fn host_state(&self) -> &dyn Any

Return a reference to the custom state attached to this instance.

pub fn memory_index(&self, memory: &VMMemoryDefinition) -> DefinedMemoryIndex

Return the memory index for the given VMMemoryDefinition in this instance.

pub fn memory_grow(
&self,
memory_index: DefinedMemoryIndex,
delta: u32
) -> Option<u32>

Grow memory in this instance by the specified amount of pages.

Returns None if memory can’t be grown by the specified amount of pages.

pub fn table_index(&self, table: &VMTableDefinition) -> DefinedTableIndex

Return the table index for the given VMTableDefinition in this instance.

pub fn table_grow(
&self,
table_index: TableIndex,
delta: u32,
init_value: TableElement
) -> Option<u32>

Grow table in this instance by the specified amount of elements.

When the table is successfully grown, returns the original size of the table.

Returns None if memory can’t be grown by the specified amount of pages or if the init_value is the incorrect table element type.

pub fn defined_table_grow(
&self,
table_index: DefinedTableIndex,
delta: u32,
init_value: TableElement
) -> Option<u32>

Grow table in this instance by the specified amount of elements.

When the table is successfully grown, returns the original size of the table.

Returns None if memory can’t be grown by the specified amount of pages or if the init_value is the incorrect table element type.

pub fn table_get(
&self,
table_index: DefinedTableIndex,
index: u32
) -> Option<TableElement>

Get table element reference.

Returns None if index is out of bounds.

pub fn table_set(
&self,
table_index: DefinedTableIndex,
index: u32,
val: TableElement
) -> Result<(), ()>

Set table element reference.

Returns an error if the index is out of bounds

pub fn defined_table_fill(
&self,
table_index: DefinedTableIndex,
dst: u32,
val: TableElement,
len: u32
) -> Result<(), Trap>

Fill a region of the table.

Returns an error if the region is out of bounds or val is not of the correct type.

pub fn get_defined_table(&self, index: DefinedTableIndex) -> &Table

Get a table defined locally within this module.

pub unsafe fn clone(&self) -> InstanceHandle

Returns a clone of this instance.

This is unsafe because the returned handle here is just a cheap clone of the internals, there’s no lifetime tracking around its validity. You’ll need to ensure that the returned handles all go out of scope at the same time.