Struct wasmtime_runtime::PoolingInstanceAllocator

pub struct PoolingInstanceAllocator { /* private fields */ }

Implements the pooling instance allocator.

This allocator internally maintains pools of instances, memories, tables, and stacks.

Note: the resource pools are manually dropped so that the fault handler terminates correctly.

Implementations

impl PoolingInstanceAllocator

pub fn new( config: &PoolingInstanceAllocatorConfig, tunables: &Tunables ) -> Result<Self>

Creates a new pooling instance allocator with the given strategy and limits.