wasmtime_runtime

Struct SharedMemory

pub struct SharedMemory(/* private fields */);

For shared memory (and only for shared memory), this lock-version restricts access when growing the memory or checking its size. This is to conform with the thread proposal: โWhen IsSharedArrayBuffer(...) is true, the return value should be the result of an atomic read-modify-write of the new size to the internal length slot.โ

Implementations

impl SharedMemory

pub fn new(plan: MemoryPlan) -> Result<Self>

Construct a new SharedMemory.

pub fn wrap( plan: &MemoryPlan, memory: Box<dyn RuntimeLinearMemory>, ty: Memory, ) -> Result<Self>

Wrap an existing Memory with the locking provided by a SharedMemory.

pub fn ty(&self) -> Memory

Return the memory type for this SharedMemory.

pub fn as_memory(self) -> Memory

Convert this shared memory into a Memory.

pub fn vmmemory_ptr(&self) -> *const VMMemoryDefinition

Return a pointer to the shared memoryโs VMMemoryDefinition.

pub fn grow( &self, delta_pages: u64, store: Option<&mut dyn Store>, ) -> Result<Option<(usize, usize)>, Error>

Same as RuntimeLinearMemory::grow, except with &self.

pub fn atomic_notify(&self, addr_index: u64, count: u32) -> Result<u32, Trap>

Implementation of memory.atomic.notify for this shared memory.

pub fn atomic_wait32( &self, addr_index: u64, expected: u32, timeout: Option<Instant>, ) -> Result<WaitResult, Trap>

Implementation of memory.atomic.wait32 for this shared memory.

pub fn atomic_wait64( &self, addr_index: u64, expected: u64, timeout: Option<Instant>, ) -> Result<WaitResult, Trap>

Implementation of memory.atomic.wait64 for this shared memory.