Struct wasmtime_runtime::VMExternRef

pub struct VMExternRef(/* private fields */);

An external reference to some opaque data.

Unlike the externref in the Wasm spec, VMExternRefs are non-nullable, and always point to a valid value. You may use Option<VMExternRef> to represent nullable references, and Option<VMExternRef> is guaranteed to have the same size and alignment as a raw pointer, with None represented with the null pointer.

VMExternRefs are reference counted, so cloning is a cheap, shallow operation. It also means they are inherently shared, so you may not get a mutable, exclusive reference to their inner contents, only a shared, immutable reference. You may use interior mutability with RefCell or Mutex to work around this restriction, if necessary.

VMExternRefs have pointer-equality semantics, not structural-equality semantics. Given two VMExternRefs a and b, a == b only if a and b point to the same allocation. a and b are considered not equal, even if a and b are two different identical copies of the same data, if they are in two different allocations. The hashing and ordering implementations also only operate on the pointer.

Example

use std::cell::RefCell;
use wasmtime_runtime::VMExternRef;

// Open a file. Wasm doesn't know about files, but we can let Wasm instances
// work with files via opaque `externref` handles.
let file = std::fs::File::create("some/file/path")?;

// Wrap the file up as an `VMExternRef` that can be passed to Wasm.
let extern_ref_to_file = unsafe { VMExternRef::new(file) };

// Get the underlying `dyn Any` via the `data` method, so you can use `Any`
// methods to perform runtime type checks and downcasts.

assert!(extern_ref_to_file.data().is::<std::fs::File>());
assert!(!extern_ref_to_file.data().is::<String>());

if let Some(mut file) = extern_ref_to_file.data().downcast_ref::<std::fs::File>() {
    use std::io::Write;
    writeln!(&mut file, "Hello, `VMExternRef`!")?;
}

Implementations

impl VMExternRef

pub unsafe fn new<T>(value: T) -> VMExternRef
where T: 'static + Any + Send + Sync,

Wrap the given value inside an VMExternRef.

Safety

Callers are responsible for ensuring that the VMExternRef is used in a thread-safe way, because the internal reference counting is not atomic.

pub unsafe fn new_with<T>(make_value: impl FnOnce() -> T) -> VMExternRef
where T: 'static + Any + Send + Sync,

Construct a new VMExternRef in place by invoking make_value.

Safety

Callers are responsible for ensuring that the VMExternRef is used in a thread-safe way, because the internal reference counting is not atomic.

pub fn as_gc_ref(&self) -> &VMGcRef

Turn this VMExternRef into a its underlying GC reference.

Unlike into_gc_ref, this does not consume and forget self. It is not safe to use from_gc_ref on pointers returned from this method; only use clone_from_gc_ref!

Nor does this method increment the reference count. You must ensure that self (or some other clone of self) stays alive until clone_from_gc_ref is called.

pub unsafe fn into_gc_ref(self) -> VMGcRef

Consume this VMExternRef into a raw, untyped pointer.

Safety

This method forgets self, so it is possible to create a leak of the underlying reference counted data if not used carefully.

Use from_gc_ref to recreate the VMExternRef.

pub unsafe fn from_gc_ref(gc_ref: VMGcRef) -> Self

Recreate a VMExternRef from a pointer returned from a previous call to {as,into}_gc_ref.

Safety

The given gc_ref must point to a valid VMExternData.

Unlike clone_from_gc_ref, this does not increment the reference count of the underlying data. It is not safe to continue to use the pointer passed to this function.

pub unsafe fn clone_from_gc_ref(gc_ref: VMGcRef) -> Self

Recreate a VMExternRef from a pointer returned from a previous call to {as,into}_gc_ref.

Safety

Wildly unsafe to use with anything other than the result of a previous {as,into}_gc_ref call!

Additionally, it is your responsibility to ensure that this raw VMExternRef is still valid, or else you could have use-after-free bugs.

pub fn as_raw(&self) -> *mut u8

Get a raw pointer for this VMExternRef.

Unlike into_raw, this does not consume and forget self. It is not safe to use from_raw on pointers returned from this method; only use clone_from_raw!

Nor does this method increment the reference count. You must ensure that self (or some other clone of self) stays alive until clone_from_raw is called.

pub unsafe fn into_raw(self) -> *mut u8

Consume this VMExternRef into a raw, untyped pointer.

Safety

This method forgets self, so it is possible to create a leak of the underlying reference counted data if not used carefully.

Use from_raw to recreate the VMExternRef.

pub unsafe fn from_raw(raw: *mut u8) -> Option<Self>

Recreate a VMExternRef from a pointer returned from a previous call to {as,into}_raw.

Safety

The given raw must point to a valid VMExternData.

Unlike clone_from_raw, this does not increment the reference count of the underlying data. It is not safe to continue to use the pointer passed to this function.

pub unsafe fn clone_from_raw(raw: *mut u8) -> Option<Self>

Recreate a VMExternRef from a pointer returned from a previous call to {as,into}_raw.

Safety

Wildly unsafe to use with anything other than the result of a previous {as,into}_raw call!

Additionally, it is your responsibility to ensure that this raw VMExternRef is still valid, or else you could have use-after-free bugs.

pub fn strong_count(&self) -> usize

Get the strong reference count for this VMExternRef.

impl VMExternRef

Methods that would normally be trait implementations, but aren’t to avoid potential footguns around VMExternRef’s pointer-equality semantics.

Note that none of these methods are on &self, they all require a fully-qualified VMExternRef::foo(my_ref) invocation.

pub fn eq(a: &Self, b: &Self) -> bool

Check whether two VMExternRefs point to the same inner allocation.

Note that this uses pointer-equality semantics, not structural-equality semantics, and so only pointers are compared, and doesn’t use any Eq or PartialEq implementation of the pointed-to values.

pub fn hash<H>(externref: &Self, hasher: &mut H)
where H: Hasher,

Hash a given VMExternRef.

Note that this just hashes the pointer to the inner value, it does not use the inner value’s Hash implementation (if any).

pub fn cmp(a: &Self, b: &Self) -> Ordering

Compare two VMExternRefs.

Note that this uses pointer-equality semantics, not structural-equality semantics, and so only pointers are compared, and doesn’t use any Cmp or PartialCmp implementation of the pointed-to values.

pub fn data(&self) -> &(dyn Any + Send + Sync)

Get the underlying host data.

pub unsafe fn data_mut(&mut self) -> &mut (dyn Any + Send + Sync)

Get a mutable borrow of the underlying host data.

Safety

It is up to the caller to ensure that there are not any other active borrows of the data.