Trait wasmtime_runtime::RuntimeLinearMemory

pub trait RuntimeLinearMemory: Send + Sync {
    fn byte_size(&self) -> usize;
    fn maximum_byte_size(&self) -> Option<usize>;
    fn grow_to(&mut self, size: usize) -> Result<()>;
    fn vmmemory(&self) -> VMMemoryDefinition;
    fn needs_init(&self) -> bool;
}

A linear memory

Required Methods

fn byte_size(&self) -> usize

Returns the number of allocated bytes.

fn maximum_byte_size(&self) -> Option<usize>

Returns the maximum number of bytes the memory can grow to. Returns None if the memory is unbounded.

fn grow_to(&mut self, size: usize) -> Result<()>

Grow memory to the specified amount of bytes.

Returns an error if memory can’t be grown by the specified amount of bytes.

fn vmmemory(&self) -> VMMemoryDefinition

Return a VMMemoryDefinition for exposing the memory to compiled wasm code.

fn needs_init(&self) -> bool

Does this memory need initialization? It may not if it already has initial contents courtesy of the MemoryImage passed to RuntimeMemoryCreator::new_memory().